Switch non-coordinate float arrays to float32 throughout _rasmap.py/raster.py
HDF result data (WSE, velocity, elevations) is already float32; the pipeline
was upcasting everything to float64 unnecessarily. Changes:

- compute_face_wss: face_value_a/b -> float32
- _compute_facepoint_wse_nb: fp_wse_at_face, buf_zs, buf_raw -> float32
- reconstruct_face_velocities: face_vel_A/B -> float32
- compute_facepoint_velocities: fp_vel_data -> float32; simplify store lines
- replace_face_velocities_sloped: replaced -> float32
- _depth_weights_for_cell: dw -> float32
- _compute_cell_pixel_weights: vel_w_batch -> float32
- compute_cell_flat_velocities: vx/vy -> float32
- _rasterize_cells_nb kernel: all per-cell working arrays -> float32
- rasterize_rasmap wrapper: sentinel arrays + all non-XY casts -> float32
- _rasterize_rasmap serial: fp_local_wse/adj, face_local_wse, cell_dw,
  nb_fp_v*, nb_face_v* -> float32
- sample_terrain_at_facepoints: grid cast -> float32, fp_elev -> float32
- raster.py: terrain_grid no longer upcast; velocity/flat-cell -> float32

XY coordinate arrays (fp_coords, cell_centers, face_midsides) remain float64.
Matrix inversions in reconstruct_face_velocities/compute_cell_flat_velocities
remain float64 internally for numerical stability.

+++ src/raspy/geo/_rasmap.py
@@ -183,9 +183,9 @@ def compute_face_wss(
 
     Returns
     -------
-    face_value_a : float64 ndarray, shape ``(n_faces,)``
+    face_value_a : float32 ndarray, shape ``(n_faces,)``
         WSE on the cellA side of each face; ``-9999`` where nodata.
-    face_value_b : float64 ndarray, shape ``(n_faces,)``
+    face_value_b : float32 ndarray, shape ``(n_faces,)``
         WSE on the cellB side of each face; ``-9999`` where nodata.
     face_hconn : uint8 ndarray, shape ``(n_faces,)``
         Hydraulic-connection classification per face (``HC_*`` constants).
@@ -195,8 +195,8 @@ def compute_face_wss(
         ``FaceValues.IsHydraulicallyConnected``.
     """
     n_faces = len(face_cell_indexes)
-    face_value_a = np.full(n_faces, _NODATA, dtype=np.float64)
-    face_value_b = np.full(n_faces, _NODATA, dtype=np.float64)
+    face_value_a = np.full(n_faces, _NODATA, dtype=np.float32)
+    face_value_b = np.full(n_faces, _NODATA, dtype=np.float32)
     face_hconn = np.zeros(n_faces, dtype=np.uint8)  # HC_NONE by default
 
     for f in prange(n_faces):
@@ -646,7 +646,7 @@ def _compute_facepoint_wse_nb(
 
     Returns
     -------
-    fp_wse_at_face : float64 ndarray, shape ``(n_faces, 2)``
+    fp_wse_at_face : float32 ndarray, shape ``(n_faces, 2)``
         Per-face arc WSE:
 
         * ``[fi, 0]`` — regression result at face ``fi`` from fpA's arc.
@@ -656,7 +656,7 @@ def _compute_facepoint_wse_nb(
     """
     n_faces = len(face_facepoint_indexes)
     n_fp = len(fp_coords)
-    fp_wse_at_face = np.full((n_faces, 2), _NODATA, dtype=np.float64)
+    fp_wse_at_face = np.full((n_faces, 2), _NODATA, dtype=np.float32)
 
     # Maximum arc buffer size: fp_count arc faces + 1 terminal face.
     # Computed once (serial reduction) so each parallel iteration can
@@ -673,8 +673,8 @@ def _compute_facepoint_wse_nb(
         # for any real HEC-RAS mesh) so allocation overhead is negligible.
         buf_xs = np.empty(max_fp_count + 1, dtype=np.float64)
         buf_ys = np.empty(max_fp_count + 1, dtype=np.float64)
-        buf_zs = np.empty(max_fp_count + 1, dtype=np.float64)
-        buf_raw = np.empty(max_fp_count, dtype=np.float64)
+        buf_zs = np.empty(max_fp_count + 1, dtype=np.float32)
+        buf_raw = np.empty(max_fp_count, dtype=np.float32)
         processed = np.zeros(max_fp_count, dtype=np.bool_)
 
         base_x = fp_coords[fp_idx, 0]
@@ -940,18 +940,18 @@ def reconstruct_face_velocities(
 
     Returns
     -------
-    face_vel_A : float64 ndarray, shape ``(n_faces, 2)``
+    face_vel_A : float32 ndarray, shape ``(n_faces, 2)``
         Full ``[Vx, Vy]`` velocity at each face reconstructed from cellA's
         C-stencil (Item1 in RASMapper terminology).
-    face_vel_B : float64 ndarray, shape ``(n_faces, 2)``
+    face_vel_B : float32 ndarray, shape ``(n_faces, 2)``
         Full ``[Vx, Vy]`` velocity at each face reconstructed from cellB's
         C-stencil (Item2 in RASMapper terminology).
         For connected faces both arrays hold the averaged value.
         Boundary faces (no cellA or cellB) fall back to ``vn * n_hat``.
     """
     n_faces = len(face_cell_indexes)
-    face_vel_A = np.zeros((n_faces, 2), dtype=np.float64)
-    face_vel_B = np.zeros((n_faces, 2), dtype=np.float64)
+    face_vel_A = np.zeros((n_faces, 2), dtype=np.float32)
+    face_vel_B = np.zeros((n_faces, 2), dtype=np.float32)
 
     for fidx in prange(n_faces):
         cellA = face_cell_indexes[fidx, 0]
@@ -1237,9 +1237,9 @@ def _compute_facepoint_velocities_nb(
 
             gj = fp_start + j
             if total_w > np.float32(1e-12):
-                # Explicit float32 division — matches C# MeshFV2D.cs result type
-                fp_vel_data[gj, 0] = float(np.float32(sum_vx / total_w))
-                fp_vel_data[gj, 1] = float(np.float32(sum_vy / total_w))
+                # float32 division — matches C# MeshFV2D.cs result type
+                fp_vel_data[gj, 0] = sum_vx / total_w
+                fp_vel_data[gj, 1] = sum_vy / total_w
 
 
 def compute_facepoint_velocities(
@@ -1340,7 +1340,7 @@ def compute_facepoint_velocities(
 
     Returns
     -------
-    fp_vel_data : float64 ndarray, shape ``(total_fp_face_entries, 2)``
+    fp_vel_data : float32 ndarray, shape ``(total_fp_face_entries, 2)``
         Flat CSR velocity store.  ``fp_vel_data[fp_face_info[fp, 0] + j]``
         gives the ``[Vx, Vy]`` arc-context velocity for facepoint ``fp``,
         local face index ``j`` (same order as the angle-sorted face ring).
@@ -1356,7 +1356,7 @@ def compute_facepoint_velocities(
     face_inv_lengths = 1.0 / np.maximum(face_lengths, 1e-12)
 
     total = int(fp_face_info[:, 1].sum())
-    fp_vel_data = np.zeros((total, 2), dtype=np.float64)
+    fp_vel_data = np.zeros((total, 2), dtype=np.float32)
 
     # np.asarray avoids an allocation when the arrays are already int64.
     _ffi64 = np.asarray(face_facepoint_indexes, dtype=np.int64)
@@ -1417,11 +1417,11 @@ def replace_face_velocities_sloped(
 
     Returns
     -------
-    replaced_face_vel : float64 ndarray, shape ``(n_faces, 2)``
+    replaced_face_vel : float32 ndarray, shape ``(n_faces, 2)``
         ``[Vx, Vy]`` sloped replacement velocity for each face.
     """
     n_faces = len(face_facepoint_indexes)
-    replaced = np.zeros((n_faces, 2), dtype=np.float64)
+    replaced = np.zeros((n_faces, 2), dtype=np.float32)
     for f in prange(n_faces):
         fpA = face_facepoint_indexes[f, 0]
         fpB = face_facepoint_indexes[f, 1]
@@ -1614,13 +1614,13 @@ def _depth_weights_for_cell(
 
     Returns
     -------
-    dw : float64 ndarray, shape ``(2 * count,)``
+    dw : float32 ndarray, shape ``(2 * count,)``
         Depth weights, minimum 0.01.  First ``count`` entries are facepoint
         depths; last ``count`` entries are face-midpoint depths.
     """
     start = int(cell_face_info[cell_idx, 0])
     count = int(cell_face_info[cell_idx, 1])
-    dw = np.full(2 * count, 0.01, dtype=np.float64)
+    dw = np.full(2 * count, np.float32(0.01), dtype=np.float32)
 
     for k in range(count):
         fi  = int(cell_face_values[start + k, 0])
@@ -1737,13 +1737,13 @@ def _compute_cell_pixel_weights(
     -------
     fw_batch : float32 ``(n_pixels, N)``
         Barycentric weights for each pixel over the N polygon corners.
-    vel_w_batch : float64 ``(n_pixels, 2*N)``
+    vel_w_batch : float32 ``(n_pixels, 2*N)``
         Donated weights for the full 2N-point stencil (N corners + N faces).
     """
     n = len(pxs)
     N = len(verts_x)
     fw_batch    = np.empty((n, N),      dtype=np.float32)
-    vel_w_batch = np.empty((n, 2 * N),  dtype=np.float64)
+    vel_w_batch = np.empty((n, 2 * N),  dtype=np.float32)
     for i in range(n):
         fw              = _barycentric_weights(pxs[i], pys[i], verts_x, verts_y)
         fw_batch[i]     = fw
@@ -2066,7 +2066,7 @@ def _sample_terrain_nb(
     """
     H, W = terrain_grid.shape
     n_fp = len(fp_coords)
-    fp_elev = np.full(n_fp, np.nan)
+    fp_elev = np.full(n_fp, np.nan, dtype=np.float32)
 
     for i in prange(n_fp):
         col_f = (fp_coords[i, 0] - c) / a - 0.5
@@ -2130,7 +2130,7 @@ def sample_terrain_at_facepoints(
     c_coef = transform.c
     d_coef = transform.e  # rasterio uses .e for the y-scale
     f_coef = transform.f
-    grid = np.asarray(terrain_grid, dtype=np.float64)
+    grid = np.asarray(terrain_grid, dtype=np.float32)
     coords = np.asarray(fp_coords, dtype=np.float64)
     return _sample_terrain_nb(coords, grid, a_coef, c_coef, d_coef, f_coef)
 
@@ -2189,8 +2189,8 @@ def compute_cell_flat_velocities(
         Velocity components.  Cells not in *flat_wet_mask* are left at 0.0.
     """
     n_cells = flat_wet_mask.shape[0]
-    vx = np.zeros(n_cells, dtype=np.float64)
-    vy = np.zeros(n_cells, dtype=np.float64)
+    vx = np.zeros(n_cells, dtype=np.float32)
+    vy = np.zeros(n_cells, dtype=np.float32)
 
     for ci in prange(n_cells):
         if not flat_wet_mask[ci]:
@@ -2268,30 +2268,30 @@ def _rasterize_cells_nb(
     cell_face_values: np.ndarray,   # (total_cf, 2) int64
     face_facepoint_indexes: np.ndarray,  # (n_faces, 2) int64
     face_cell_indexes: np.ndarray,  # (n_faces, 2) int64
-    face_min_elev: np.ndarray,      # (n_faces,) float64
+    face_min_elev: np.ndarray,      # (n_faces,) float32
     fp_coords: np.ndarray,          # (n_fp, 2) float64
     # WSE
-    cell_wse: np.ndarray,           # (n_cells,) float64
-    face_value_a: np.ndarray,       # (n_faces,) float64
-    face_value_b: np.ndarray,       # (n_faces,) float64
-    fp_wse: np.ndarray,             # (n_faces, 2) float64  or  (0, 2) sentinel
+    cell_wse: np.ndarray,           # (n_cells,) float32
+    face_value_a: np.ndarray,       # (n_faces,) float32
+    face_value_b: np.ndarray,       # (n_faces,) float32
+    fp_wse: np.ndarray,             # (n_faces, 2) float32  or  (0, 2) sentinel
     has_fp_wse: bool,
     # Terrain
-    terrain_grid: np.ndarray,       # (H_t, W_t) float64  or  (1, 1) sentinel
+    terrain_grid: np.ndarray,       # (H_t, W_t) float32  or  (1, 1) sentinel
     has_terrain: bool,
     # Velocity data
-    fp_vel_data: np.ndarray,        # (total, 2) float64  or  (0, 2) sentinel
+    fp_vel_data: np.ndarray,        # (total, 2) float32  or  (0, 2) sentinel
     fp_face_info_arr: np.ndarray,   # (n_fp, 2) int64  or  (0, 2) sentinel
     face_fp_local_idx: np.ndarray,  # (n_faces, 2) int32  or  (0, 2) sentinel
-    face_vel_A: np.ndarray,         # (n_faces, 2) float64  or  (0, 2) sentinel
-    face_vel_B: np.ndarray,         # (n_faces, 2) float64  or  (0, 2) sentinel
+    face_vel_A: np.ndarray,         # (n_faces, 2) float32  or  (0, 2) sentinel
+    face_vel_B: np.ndarray,         # (n_faces, 2) float32  or  (0, 2) sentinel
     has_vel_data: bool,
     # Flat-cell velocity
-    flat_cell_vx: np.ndarray,       # (n_cells,) float64  or  (0,) sentinel
-    flat_cell_vy: np.ndarray,       # (n_cells,) float64  or  (0,) sentinel
+    flat_cell_vx: np.ndarray,       # (n_cells,) float32  or  (0,) sentinel
+    flat_cell_vy: np.ndarray,       # (n_cells,) float32  or  (0,) sentinel
     has_flat_vel: bool,
     # Facepoint terrain elevation (depth-weight rebalancing)
-    fp_elev: np.ndarray,            # (n_fp,) float64  or  (0,) sentinel
+    fp_elev: np.ndarray,            # (n_fp,) float32  or  (0,) sentinel
     has_fp_elev: bool,
     face_hconn: np.ndarray,         # (n_faces,) uint8
     # Render options
@@ -2351,9 +2351,9 @@ def _rasterize_cells_nb(
             verts_y[k]  = fp_coords[fp_i, 1]
 
         # ---- Per-cell WSE arrays --------------------------------------------
-        fp_local_wse     = np.full(N, _NODATA, dtype=np.float64)
-        fp_local_wse_adj = np.full(N, _NODATA, dtype=np.float64)
-        face_local_wse   = np.full(N, _NODATA, dtype=np.float64)
+        fp_local_wse     = np.full(N, np.float32(_NODATA), dtype=np.float32)
+        fp_local_wse_adj = np.full(N, np.float32(_NODATA), dtype=np.float32)
+        face_local_wse   = np.full(N, np.float32(_NODATA), dtype=np.float32)
         use_sloped  = False
         has_face_wse = False
 
@@ -2400,7 +2400,7 @@ def _rasterize_cells_nb(
                         face_local_wse[k] = cws
 
         # ---- Depth weights --------------------------------------------------
-        cell_dw = np.empty(0, dtype=np.float64)
+        cell_dw = np.empty(0, dtype=np.float32)
         if use_sloped and with_faces and use_depth_weights and has_fp_elev:
             cell_dw = _depth_weights_for_cell(
                 cell_idx, cell_face_info, cell_face_values,
@@ -2409,15 +2409,15 @@ def _rasterize_cells_nb(
             )
 
         # ---- Per-cell velocity arrays ---------------------------------------
-        nb_fp_vx   = np.empty(0, dtype=np.float64)
-        nb_fp_vy   = np.empty(0, dtype=np.float64)
-        nb_face_vx = np.empty(0, dtype=np.float64)
-        nb_face_vy = np.empty(0, dtype=np.float64)
+        nb_fp_vx   = np.empty(0, dtype=np.float32)
+        nb_fp_vy   = np.empty(0, dtype=np.float32)
+        nb_face_vx = np.empty(0, dtype=np.float32)
+        nb_face_vy = np.empty(0, dtype=np.float32)
         has_vel_arrays = False
 
         if (variable_flag == _VAR_SPEED or variable_flag == _VAR_VELOCITY) and has_vel_data:
-            nb_fp_vx = np.zeros(N, dtype=np.float64)
-            nb_fp_vy = np.zeros(N, dtype=np.float64)
+            nb_fp_vx = np.zeros(N, dtype=np.float32)
+            nb_fp_vy = np.zeros(N, dtype=np.float32)
             for i in range(N):
                 fp_i = verts_fp[i]
                 fi   = face_indices[i]
@@ -2428,8 +2428,8 @@ def _rasterize_cells_nb(
                     nb_fp_vx[i] = fp_vel_data[offset, 0]
                     nb_fp_vy[i] = fp_vel_data[offset, 1]
 
-            nb_face_vx = np.zeros(N, dtype=np.float64)
-            nb_face_vy = np.zeros(N, dtype=np.float64)
+            nb_face_vx = np.zeros(N, dtype=np.float32)
+            nb_face_vy = np.zeros(N, dtype=np.float32)
             for j in range(N):
                 fi  = face_indices[j]
                 ori = face_orients[j]
@@ -2656,11 +2656,11 @@ def rasterize_rasmap(
     pix_ys   = ys[sort_order]
 
     # Sentinel arrays for optional inputs (Numba requires consistent types)
-    _e2f  = np.empty((0, 2), dtype=np.float64)   # (0,2) float64 sentinel
+    _e2f  = np.empty((0, 2), dtype=np.float32)   # (0,2) float32 sentinel
     _e2i  = np.empty((0, 2), dtype=np.int64)      # (0,2) int64 sentinel
     _e2i32= np.empty((0, 2), dtype=np.int32)      # (0,2) int32 sentinel
-    _e1f  = np.empty(0,      dtype=np.float64)    # (0,)  float64 sentinel
-    _terr = np.empty((1, 1), dtype=np.float64)    # terrain sentinel (non-zero shape)
+    _e1f  = np.empty(0,      dtype=np.float32)    # (0,)  float32 sentinel
+    _terr = np.empty((1, 1), dtype=np.float32)    # terrain sentinel (non-zero shape)
 
     _rasterize_cells_nb(
         pix_rows, pix_cols, pix_xs, pix_ys,
@@ -2672,30 +2672,30 @@ def rasterize_rasmap(
         np.asarray(cell_face_values,        dtype=np.int64),
         np.asarray(face_facepoint_indexes,  dtype=np.int64),
         np.asarray(face_cell_indexes,       dtype=np.int64),
-        np.asarray(face_min_elev,           dtype=np.float64),
+        np.asarray(face_min_elev,           dtype=np.float32),
         np.asarray(fp_coords,               dtype=np.float64),
         # WSE
-        np.asarray(cell_wse,     dtype=np.float64),
-        np.asarray(face_value_a, dtype=np.float64),
-        np.asarray(face_value_b, dtype=np.float64),
-        np.asarray(fp_wse,       dtype=np.float64) if fp_wse is not None else _e2f,
+        np.asarray(cell_wse,     dtype=np.float32),
+        np.asarray(face_value_a, dtype=np.float32),
+        np.asarray(face_value_b, dtype=np.float32),
+        np.asarray(fp_wse,       dtype=np.float32) if fp_wse is not None else _e2f,
         fp_wse is not None,
         # Terrain
-        np.asarray(terrain_grid, dtype=np.float64) if terrain_grid is not None else _terr,
+        np.asarray(terrain_grid, dtype=np.float32) if terrain_grid is not None else _terr,
         terrain_grid is not None,
         # Velocity
-        np.asarray(fp_vel_data,       dtype=np.float64) if fp_vel_data       is not None else _e2f,
+        np.asarray(fp_vel_data,       dtype=np.float32) if fp_vel_data       is not None else _e2f,
         np.asarray(fp_face_info,      dtype=np.int64)   if fp_face_info      is not None else _e2i,
         np.asarray(face_fp_local_idx, dtype=np.int32)   if face_fp_local_idx is not None else _e2i32,
-        np.asarray(face_vel_A,        dtype=np.float64) if face_vel_A        is not None else _e2f,
-        np.asarray(face_vel_B,        dtype=np.float64) if face_vel_B        is not None else _e2f,
+        np.asarray(face_vel_A,        dtype=np.float32) if face_vel_A        is not None else _e2f,
+        np.asarray(face_vel_B,        dtype=np.float32) if face_vel_B        is not None else _e2f,
         fp_vel_data is not None,
         # Flat-cell velocity
-        np.asarray(flat_cell_vx, dtype=np.float64) if flat_cell_vx is not None else _e1f,
-        np.asarray(flat_cell_vy, dtype=np.float64) if flat_cell_vy is not None else _e1f,
+        np.asarray(flat_cell_vx, dtype=np.float32) if flat_cell_vx is not None else _e1f,
+        np.asarray(flat_cell_vy, dtype=np.float32) if flat_cell_vy is not None else _e1f,
         flat_cell_vx is not None,
         # Facepoint elevation
-        np.asarray(fp_elev, dtype=np.float64) if fp_elev is not None else _e1f,
+        np.asarray(fp_elev, dtype=np.float32) if fp_elev is not None else _e1f,
         fp_elev is not None,
         np.asarray(face_hconn, dtype=np.uint8),
         # Options
@@ -3003,7 +3003,7 @@ def _rasterize_rasmap(
             # fp_wse shape is (n_faces, 2): col 0 = fpA arc, col 1 = fpB arc.
             # ori > 0 → this cell is cellA → corner facepoint is fpA → col 0.
             # ori < 0 → this cell is cellB → corner facepoint is fpB → col 1.
-            fp_local_wse = fp_wse[face_indices, _fp_col].astype(np.float64)
+            fp_local_wse = fp_wse[face_indices, _fp_col].astype(np.float32)
             if with_faces:
                 # Face-midpoint WSEs: select the side of the face that belongs
                 # to this cell.  face_value_a corresponds to cellA (column 0
@@ -3011,7 +3011,7 @@ def _rasterize_rasmap(
                 _is_cellA      = cell_idx == face_cell_indexes[face_indices, 0]
                 face_local_wse = np.where(
                     _is_cellA, face_value_a[face_indices], face_value_b[face_indices]
-                ).astype(np.float64)
+                ).astype(np.float32)
                 # Valid when at least one corner OR face-midpoint has a real WSE.
                 has_valid_wse = (fp_local_wse != _NODATA).any() or (face_local_wse != _NODATA).any()
             else:
@@ -3067,7 +3067,7 @@ def _rasterize_rasmap(
         # Only computed when use_depth_weights=True (opt-in, default off).
         # Requires fp_elev; silently skipped if unavailable (cell_dw length-0
         # causes _pixel_wse_sloped to use plain donated weights instead).
-        cell_dw: np.ndarray = np.empty(0, dtype=np.float64)
+        cell_dw: np.ndarray = np.empty(0, dtype=np.float32)
         if use_sloped and with_faces and use_depth_weights and fp_elev is not None:
             cell_dw = _depth_weights_for_cell(
                 cell_idx, cell_face_info, cell_face_values,
@@ -3092,8 +3092,8 @@ def _rasterize_rasmap(
             # Matches C# GetLocalFacepointValues → fpVelocityRing[fPPrev]
             # .Velocity[...].  Do NOT use replaced_face_vel (face-averaged
             # facepoint velocity) — that was wrong for the pixel stencil.
-            nb_fp_vx = np.zeros(N, dtype=np.float64)
-            nb_fp_vy = np.zeros(N, dtype=np.float64)
+            nb_fp_vx = np.zeros(N, dtype=np.float32)
+            nb_fp_vy = np.zeros(N, dtype=np.float32)
             for i in range(N):
                 fp_i = verts_fp[i]
                 fi   = face_indices[i]
@@ -3110,8 +3110,8 @@ def _rasterize_rasmap(
             # Face-midpoint velocities: select the cell-side value from the
             # two face velocity arrays (face_vel_A for cellA, face_vel_B for
             # cellB), matching the same cellA/cellB logic as the WSE arrays.
-            nb_face_vx = np.zeros(N, dtype=np.float64)
-            nb_face_vy = np.zeros(N, dtype=np.float64)
+            nb_face_vx = np.zeros(N, dtype=np.float32)
+            nb_face_vy = np.zeros(N, dtype=np.float32)
             for j in range(N):
                 fi  = face_indices[j]
                 ori = face_orients[j]
